pyproject.toml
[project]
name = "resume-tailor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.48.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "agents>=0.1.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "openai-agents>=0.2.9",
    "pypdf>=6.0.0",
]
